pyproject.toml
[tool.poetry]
name = "tui-network"
version = "0.1.0"
description = "A Python based TUI for managing wifi on Linux"
authors = ["Matteo Ferrini <mimmi.ferrini@gmail.com>"]
readme = "README.md"
include = [
    "tui_network/assets"
]

[tool.poetry.scripts]
tui-network = "tui_network.main:app"

[tool.poetry.dependencies]
python = "^3.11"
textual = "^0.89.1"
dbus-python = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
